# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v2/internal/users": {
            "get": {
                "description": "Get all information needed about user in list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users-internal"
                ],
                "summary": "Get All Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "can only be: communityId, name, email, phoneNumber",
                        "name": "searchBy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "inputted search based on searchBy",
                        "name": "search",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination",
                        "name": "cursor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pagination direction - prev or next",
                        "name": "direction",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "how many data that user want to load",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "filter by campus",
                        "name": "campusCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "filter by cool",
                        "name": "coolId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "filter by department",
                        "name": "departmentCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Pagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetAllUserCursorResponse"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/definitions/models.CursorInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Events based on User Roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get All Events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.List"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetAllEventsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/events/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Event and Instances by Event Code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event by Event Code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "object that needs to be added",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ListWithDetail"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetInstancesByEventCodeResponse"
                                            }
                                        },
                                        "details": {
                                            "$ref": "#/definitions/models.GetEventByCodeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v2/internal/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create event with the instances/sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create Event",
                "parameters": [
                    {
                        "description": "User object that needs to be added",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.CreateEventResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "instances": {
                                            "$ref": "#/definitions/models.CreateInstanceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/internal/events/{eventCode}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "For Internal Purposes Only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event and Sessions by Event Code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "object that needs to be added",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ListWithDetail"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetInstanceSummaryResponse"
                                            }
                                        },
                                        "details": {
                                            "$ref": "#/definitions/models.GetEventSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all information needed about user by community id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User By Access Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.GetOneByCommunityIdResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.RoleResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create user for all",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "User object that needs to be added",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/users/{communityId}": {
            "get": {
                "description": "Get all information needed about user by community id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User By Community ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "object that needs to be added",
                        "name": "communityId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mandatory header to access endpoint",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response indicates that the request succeeded and the resources has been fetched and transmitted in the message body",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.GetOneByCommunityIdResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.RoleResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error. This can happen if there is an error validation while create account",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/definitions/models.ErrorValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CursorInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "parameters": {},
                "previous": {
                    "type": "string"
                },
                "totalData": {
                    "type": "integer"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ErrorValidateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "accountNumber_required"
                },
                "field": {
                    "type": "string",
                    "example": "MISSING_FIELD"
                },
                "message": {
                    "type": "string",
                    "example": "field is missing"
                }
            }
        },
...
        "models.GetAllUserCursorResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "campusCode": {
                    "type": "string"
                },
                "campusName": {
                    "type": "string"
                },
                "communityId": {
                    "type": "string"
                },
                "coolId": {
                    "type": "integer"
                },
                "coolName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "departmentCode": {
                    "type": "string"
                },
                "departmentName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "isBaptized": {
                    "type": "boolean"
                },
                "isKom100": {
                    "type": "boolean"
                },
                "jemaatId": {
                    "type": "string"
                },
                "kkjNumber": {
                    "type": "string"
                },
                "maritalStatus": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "placeOfBirth": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.List": {
            "type": "object",
            "properties": {
                "data": {},
                "totalRows": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "example": "collection"
                }
            }
        },
        "models.ListWithDetail": {
            "type": "object",
            "properties": {
                "data": {},
                "details": {},
                "totalRows": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "example": "collection"
                }
            }
        },
...
        "models.Pagination": {
            "type": "object",
            "properties": {
                "data": {},
                "details": {},
                "pagination": {},
                "type": {
                    "type": "string",
                    "example": "collection"
                }
            }
        },
...
        "models.RoleResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "View specifically for event"
                },
                "role": {
                    "type": "string",
                    "example": "event-view-volunteer"
                },
                "type": {
                    "type": "string",
                    "example": "role"
                }
            }
        },
...
    }
}